+++ test/index.ts
@@ -10,3 +10,5 @@ import './theme/apply';
 import './storage/ThemeOption';
 import './pages/options/hooks/useTheme';
 import './theme/mirror';
+import './pages/notifications/components/Alert/Alert';
+import './pages/options/components/Options/Modal/Modal';